requirements.txt
# Telegram Bot Framework
python-telegram-bot[job-queue]>=20.7

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Redis Client
redis>=5.0.1

# HTTP Client
httpx>=0.26.0

# Configuration
python-decouple>=3.8

# Logging
structlog>=24.1.0

# Utilities
python-dateutil>=2.8.2

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
